CreateStatusbar and SetStatusText moved to MapFrameBase
